Add missing check for positional array aggregate
Positional array aggregate should have index checks
for every single of their implicit indices. Those
were not generated because they are in general
redundant with the checks on bound, but this is only
correct with indexes that are within signed integer
types, and without an 'others' choice.

For modular types, the bound computation could
wrapround, making the implicit indices fall out
of the range because the range becomes empty.
We generate an additional check for this case.

For positional aggregate with an 'others' choice,
there could be more positional association than
allowed by the index bounds, which are taken from
the context in that case. We generate an additional
check for this case. In general, the two checks
correspond to a range check for the implicit index
of the last positional association.

A similar case arise for empty array aggregates, with
wrapround ocurring in the other direction. This
was already prevented by pre-existing checks generated for
toplevel empty array aggregates, but not for empty
subaggregates. We move the generation of those
checks to Insert_Checks main recursion to generate
checks correctly in that case.