Improve labeling global objects as Output instead of In_Out
When deciding whether a global object is Output or In_Out we should
ignore uses of array bounds and discriminants. (This is essentially the
same logic as in Find_Ineffective_Imports_And_Unused_Objects and
Compute_Dependency_Relation, but it is quite hard to refactor the code
and have it implemented only once.)

[changelog]
* src/flow/flow-slice.adb
(Compute_Globals): Improve handling of array bounds and discriminants.
* testsuite/gnatprove/tests/292__flow_global_discriminant/test.out
Update dedicated test with discriminant; spurious check is gone.
* testsuite/gnatprove/tests/O210-006__use_generic/test.out
* testsuite/gnatprove/tests/O210-007__use_generic/test.out
Update existing tests; spurious checks are gone.