Fix computation of sanitized name to use for gprbuild invocation
We are computing the unit for a given argument on the command line, then
determine the file name of the "main part" to give to gprbuild (this is
to run gprbuild with "a.adb", even if the user provided "a.ads").
However, to get the filename we called "Base_Name", which apparently
returns the name without extension. This seemed to work by accident in
most cases, but breaks in some corner cases such as separates with
naming schemes.

* configuration.adb
(Sanitize_File_List): Use the simple name (with extension) instead of
the basename (without extension)