Add support for functions with side-effects in marking
Allow functions with outputs when they are identified as functions
with side-effects.
Check that functions with side-effects are only called as the expression
of an assignment statement.
Allow contracts Always_Terminates and Exceptional_Cases on
functions with side-effects.
Update the explain code on global outputs in functions without side-effects.

Add test. Update existing tests.