Add checks for conversions in LHS of assignments
Add systematic checks on conversions in LHS of assignments. For
constraint checks on the object itself, look through conversions to
get the object which is actually modified to decide whether or not to
emit the check.

[changelog]
* src/why/gnat2why-expr.adb
(Expected_Type_Of_Prefix): Avoid crash on components of view
 conversions.
(Shift_Rvalue): Systematically introduce checks for conversions in the
 LHS of assignments and copy back of parameters.
(Transform_Assignment_Statement): For length checks and discriminant
 checks, look through conversions to get the object which is actually
 modified to decide whether or not to emit the check.

* tests/*/test.out
Duplicated discriminant checks.

* tests/360__constraints_check_on_conv_in_LHS
New test.