Add support for functions with side-effects in proof
Functions with side-effects are handled like procedures in proof.
Extract the translation of procedures and entries into a new procedure
Transform_Procedural_Subprogram_Call, which is called for the
translation of functions with side-effects.

This patch requires that some checks from flow are bypassed currently
for functions with side-effects. This will be replaced by support in
flow in another patch.

Add explain code for calls to functions with side-effects outside of
assignments.

Add test. Minor test update on lost counterexample.