Add test with No_Return and no body
This test is about a procedure with No_Return and no body (unit Noret),
whose contract is picked in phase 1 and causes the check of
Always_Terminates on a remote caller to fail (unit Main).

We can't just have a direct call from Main to Noret, because it would
be pruned from the CFG (just like it happens for the intermediate unit
Proxy).