Forbid functions with side-effects in primitive equality
Due to composition of primitive equality operators for records,
and to implicit calls to primitive equality for such types in
membership tests, we already forbid in SPARK user-defined equality
for record types that have a non-null Global contract. Extend
this restricton to also forbid such functions to be volatile or
have side-effects, and forbid a ghost function to apply to non-ghost
record type.

Update the SPARK RM.
Add corresponding checks in marking.
Add a test.