Emit checks for user-defined primitive equalities
Primitive equalities of records and limited types will be used
implicitly in membership tests and predefined equality of composite
type. Emit checks to ensure that they can be called in any context.
Remove the detection of preconditions on primitive equality in marking.
Also check the primitive equality on membership tests.