Support bounded predefined maps and sets
If the return type of the (optional) length function of a type with
predefined set or map aggregate is a scalar type, add an assumption
when verifying Add that the length is strictly less than the maximum
length. The assumption is translated as an additional check on the
precondition of aggregate functions.