Revert update to test output after applying a proper fix
A proper fix to a recent test output instability was done in
Mark_Subprogram_Specification.

The diff was caused by an arbitrary order of iteration over a hashed
container, where the hash depends on the integer value of a Node_Id and
can change when the (implicitly) withed runtime unit is modified.