Coverage is not a project file mode anymore
Previusly, Coverage was one of the 4 modes of compilation for gnat2why.
Devel, Debug and Production are the other three, with Devel and Debug
being different forms of a run with assertions. This meant that one had
to choose between coverage and assertions.

Changing the project file so that Coverage is an independent setting (it
simply forces "-O0").

* gnat2why/Makefile:
  pass coverage options via new variable COVERAGEARGS instead of the
  previously common var GPRARGS
* gnat2why/gnat2why_gnat.gpr:
  make Coverage a separate setting