Support supplying a capacity function for all kinds of container
The capacity function shall have a container parameter iff the empty
function takes the capacity as a parameter (the capacity is specific
to each container object and not global for the container type).
If the capacity is provided, it will be used to bound the number of
elements in the container. The return subtype of the Length function is
still used as a default for predefined sets and maps if no capacity
function is provided. Supplying a capacity function on a predefined
set or map container without a length function is allowed.
Using an empty function with a capacity parameter or a constant instead
of a function are now supported.