Add tests with termination of imported subprograms with No_Return
All tests include aspect Always_Terminates which should be flagged as
wrong, because of calls to imported subprograms with No_Return.

In the original reproducer the non-returning subprogram has no global
outputs and currently is pruned, but in the other reproducers the
No_Return aspect is ignored.

[changelog]
* testsuite/gnatprove/tests/418__flow_terminates_no_return_import/
Add; missing checks.