Don't descend into existential quantifiers for splitting of VCs
The code for descending into subexpressions for the purpose of splitting
VCs was descending into all quantifiers. But it doesn't make sense for
"for some quantifiers. Refining the corresponding condition.

This has the side effect of disabling inlining for expression functions
which have a "for some" at the top level, as such functions won't
contain pretty-printing labels, and only such functions will be inlined.
(The inlining discussed here is the inlining that happens in Why3
transformations, for the purpose of providing better identification of
unproved parts of an assertion).

* gnat2why-expr.adb
(Is_Terminal_Node): Refine condition to return false for "for some"
  quantified expressions