Assume value of tag of the result of prog functions
For functions with tagged results, assume the value of the tag of the
result in the post of the program function. This should not change
much as this assumption is already done on the logic function. It
should still be useful for functions without an associated logic
function (a volatile function or a function with side-effects).