Minor refactoring for checks on access-to-subprograms
Pass an additional As_Closure parameter to decide if the value of
additional global inputs in the source should be taken from the context
or if all possible values should be considered.
As_Closure is True for specialized actuals. All global accesses are
constant and can be taken from the current context.