Support Hide_Info and Unhide_Info
The annotations can be located either at the top of a body or just
after a body or specification. The only kind of information that can
be removed currently is an expression function body but this could be
extended. A Hide_Info annotation located on its argument entity is
considered to provide a default. Conflicting Hide_Info and Unhide_Info
annotations are only allowed to override the default for the
verification of the entity itself.