Do not include the defining axiom of recursive expression functions
We do not include the defining axiom of recursive expression functions
when verifying entities which are cyclic for proof.
We only do that for expression functions which have a numeric
subprogram variant. The axiom has the form `f params = expr` which is
always sound unless `expr` depends on `f params`, which should not be
possible if `f` is known to terminate for non-proof-related reasons.

[changelog]
* src/why/gnat2why-subprograms.adb
(Translate_Expression_Function_Body): Register potentially proof
 cyclic functions and warn the user.

* src/why/why-atree-modules.adb
(Mutually_Recursive_Modules): Exclude the defining axiom of expression
 functions if necesary.

* tests/453__recursive_expr_fun
File test.

* tests/W130-026__annotate/main.adb
* tests/R509-011__higher_order/test_higher_order.ad*
* tests/U908-033__hidden_pointers/example_recursive.adb
Introduce a separated procedure to workaround imprecision in proof
cycles with the enclosing unit.

* tests/Q615-017__factorial/test.out
The overflow checks are now unprovable as we do not have the axiom on
the recursive expression function.

* tests/*/test.out
New info messages. Minor diffs is provability.