Always put the post axiom of functions in its own module
Now that the completion of functions has been splitted in multiple
modules, it does not make much sense to keep a special module for post
axioms of recursive functions. Instead, we always store the post axiom
in its own module and exclude this module in case of proof cyclic
dependencies.