Hide expression function body
Hiding the body of an expression functions is done in two ways. The
module for the defining axiom is not included, and a specific program
function without the definition in its post is introduced.
Before generating a module for VCs, we take into account the potential
Hide or Unhide annotations. The module for the defining axiom is
included/excluded from the dependency graph and the appropriate program
function symbol is stored in the symbol table.