Reject scalar type with too large size in UC
If a scalar type has a too large value of Size, it is unfit as
source of Unchecked_Conversion, as GNAT does not guarantee zeroing
out extra bits when writing the component. Adapt SPARK RM rule
and add the corresponding detection in Type_Is_Suitable_For_UC.

Patch for 24.1 uses RM_Size instead of Get_Attribute_Value which
is not available on 24.1.

Add test.