Introduce program function for array aggregates
Use an instance of the defining axiom as a post. Guards for soundness
are not necessary as the postcondition is only assumed in the context
of use. This changes results in minor improvements and regressions,
both because of proof instability and depending on whether or not the
value of the aggregate is actually useful.