Remove reading and writing of unused "session_dir" field
This field was present in gnatwhy3 output and part of the information
for each check in gnat2why JSON output (.spark file). This information
is not needed anymore and is now removed from both.

The information was previously used to find the session for a given
check, for manual proof inside GS.