Avoid the "Continuation" argument of various output routines
Some message output routines have a Boolean argument "Continuation",
which can be set to true to indicate that the message to be output is
part of the previous message, as a continuation message.

We remove this Boolean argument and replace it by an argument which
explicitly lists the continuation messages of the main message. This
design has the advantage that no duplicate/spurious information is
passed (continuation messages should not have severity, a fix message,
etc).

Continuation messages are not printed for justified messages. Also,
continuation messages cannot be justified individually, only the main
message can be.

Comments on test updates:
- RM references only printed for main message
- some "explain" or other continuation messages were incorrectly
  suppressed previously
- continuation messages were counted in summary