Simple sanity check test for SPARK
We add a new test whose intention is to act as an automated sanity check
that a built SPARK is OK. The test currently outputs the version as well
as runs a copy of the "database" test for each prover. The output should
be empty (everything proved).

Prover versions can differ for different SPARK installs, but the output
diff should be easy to check.