Deconstruct Get_Range_Check_Info
Only use Get_Range_Check_Info determine the check name (index,
overflow, length, or range check) from the Ada node when introducing a
scalar conversion. Until now, the Ada node was used also to determine
the bounds of the range check to be applied if any as well as special
cases (for parameters of 'Succ and 'Pred in particular). Now scalar
conversions introduce range checks depending only on the type they are
converting too. It uses the Ada node of the target of the conversion.
Care should be taken to only convert to predefined integer types (int
and BVs) if the range is known to be preserved. 'Succ and 'Pred are now
handled specifically.

Some checks are removed/added by this patch. They seem to all be either
spurious or redundant.