Cleanup recursive flow query for getting types
Routine Get_Type, when called with a subexpression, was first finding a
type and then, when called recursively, checked if this is indeed a type
or an abstract state.

Now once it finds a type, it finds its most appropriate view in a loop
without further checks.

[changelog]
* src/flow/flow-control_flow_graph.adb
(Get_Array_Index): Don't call Etype, as it will be called by Get_Type
anyway.
* src/flow/flow_utility.adb
(Get_Type): Rewrite without recursion; add assertions and comments.