Split invariant checks on subprograms and add continuations
We factorize the computation of invariants for inputs/outputs of the
subprogram in a generic procedure. It is reused to compute the
invariants as a whole as a predicate and to generate separate checks.
Continuations are added to specify both when the invariant check occurs
(before the call or at the end of the subprogram) and the relevant
input/output.