Recognize statically null constants and disable checks on them
A constant is considered to be statically reclaimed if either it is
annotated as such, or its definition is visible in SPARK and is either
null or statically reclaimed.
These constants are not subjected to ownership and they can safely be
used in a comparison even if it involves pointer equality.