Change switch --proof-warnings to have value on or off
It is useful in some cases to disable proof warnings, for example
in the benchmark testsuite, so that we don't generate VCs for those.
To that end, we change the switch --proof-warnings to require an
argument of "on" or "off". This is an incompatible change, that
requires users to adopt the new syntax.

[changelog]
* docs/ug/en: Update UG.
* docs/ug/table_generator/table_generator.adb: Same.
* share/spark/help.txt: Update help text.
* src/gnatprove/configuration.adb (Parse_Switches: Update switch.
  (File_Specific_Postprocess): Take new value into account.
  (Read_Command_Line): Set new value based on switch.
* src/gnatprove/configuration.ads: Update switch. Add global value.
* testsuite/gnatprove/lib/python/test_support.py: Set switch for
  benchmark testing.
* testsuite/gnatprove/tests: Update use of switch.