Support new Mutable_In_Parameters annotation
This annotation can appear just after an entry or subprogram with side
effects. It states that *in* parameter of a particular type should be
considered mutable by the tool. This only works for private type which
are ultimately either access-to-variable types or out of SPARK. This is
meant to be a workaround for third party libraries which cannot be
changed to use mode *in out* instead of *in*.

This patch adds a restriction to Is_Constant_In_SPARK which can now
only be called on formal parameters if their enclosing subprogram has
been marked. It is necessary to ensure that all annotations have been
processed for the subprogram.