Clean-up handling of visibility in proof
Proof should look into private parts of packages used from the analyzed
code but not inside their bodies. In particular, the refined
postconditions of subprograms and the bodies of expression functions
located in package bodies should only be included when the analyzed
entity has visibility over the enclosing package body. For refined
postconditions, this was implemented by having specific versions of the
program and logic symbols using either the refined or standard posts.
It was not attempted for expression functions unless they had a refined
post.
Having a different version of the called subprogram works well for
program symbols, but not for logic symbols which can occur in other
subprogram contracts which might not have the same visibility as the
analyzed entity. Instead, we introduce a single function symbol but
two post axioms, one with the refined postcondition, and one with the
normal one. The proper axiom to use is chosen by including the
appropriate module when doing the closure of included modules. This is
done using a mask when computing the closure based on the visibility of
the analyzed entity. This mask replaces axiom modules in the module
dependency map.
We use the same mechanism to handle expression functions whose body
is deferred to the enclosing package's body. We create a specific
program symbol for the refined version of the function, even if it has
no refined post. We use the same refinement mask to exclude the
defining axioms of these expression functions when they are not
visible. We also add a check that we never attempt to hide/unhide the
body of an expression function using Hide_Info if it is not visible.
So that the overriding used to handle Hide_Info continues working, we
store the identifiers for Refine and Dispatch function symbols in the
symbol table along with the standard ones.