Support for new --limit-lines option
This option allows users to specify a list of lines (or checks) in a
file.

This option needs to be passed to gnat2why and gnatwhy3. Given how the
information passing works, it was easiest to pass the list of lines to
gnat2why, and the file itself to gnatwhy3.