Use Hide_Info to hide private part in user code
Do not pull full views of types and constants and expression functions
in private parts of withed units if they have the annotation. Packages
which are not fully visible cannot be annotated this way for now
(unless they are instances of generic units, in which case the
annotation is ignored). The private part of parents remains visible in
child units, and the private part of nested packages remains visible
in the enclosing unit.