Take into account ghost predicates to reject duplicated predicates
We detect and reject private types with a full view not in SPARK if
they have a predicate both on the partial and on the full view as
there is no way to make the difference between the two in the
predicate function. Ghost predicates are now taken into account
during this detection.