Reject instances of generics from scopes containing invariants
Reject instances of generics declared in a package P which declares
a private type with a type invariant (or in generic child units of P)
if they are not themselves (the instances) inside P (or in generic
child units). This ensures that the current handling of type invariants
in GNATprove based only on scopes is valid.