Support ownership and predefined_equality annotations on hidden types
Types with ownership or a restricted predefined equality can now be
declared in hidden private parts provided they have a matching
ownership or predefined equality annotation. The fact that the
annotations on the partial view are consistent with the full view is
mostly checked in marking. However, the consistency of the reclamation
entities is deferred to proof and associated to a new kind of checks on
types.