Only reject non-conforming hidden types if they are full views
It is OK to hide a non-conforming type (for example a type with
ownership) if it is not a full view of a private type visible from the
outside of the package. Raise violations on visible partial views if
the full view is hidden and non-conforming instead.
Reject Hide_Info on private parts of packages which are not in SPARK.
This avoids going to analysis with types whose full view is in SPARK
but whose partial view was rejected.