Create body for null procedures in remaining cases
In some cases, checks on null procedures are not emitted because no
body is generated for them.

The GNAT frontend does not create a body for null procedures when
they are not completions of an existing declaration. Doing it in the frontend
is impractical as this interferes with the checking of freezing rules, so
would require the same machinery to disable freezing checks as done for
generation of bodies for expression functions.

Instead, generate the bodies in SPARK rewriting phase, to avoid such
interactions.

Add tests.

[changelog]
* src/flow/flow-analysis.adb (Is_Param_Of_Null_Procedure): New query.
  (Warn_On_Subprogram_With_No_Effect): Special case for null procedures.
  (Find_Unwritten_Exports): Same.
  (Warn_On_Unused_Objects): Same.
* src/spark/spark_rewrite.adb (Rewrite_Null_Procedure): New procedure.
  (Rewrite_Node): Call the new procedure.
* src/spark/spark_util-subprograms.ads (Was_Null_Procedure): Recognize
  generated body for null procedures.
* testsuite/gnatprove/tests/595__null_procedure: Add test.
* testsuite/gnatprove/tests: Improved outputs. Fixes missing checks on out
  parameters of null procedures. Removes spurious warnings on null
  procedures.