Do not consider out-of-bounds values in arrays
The fact that only the values inside the bounds of an array matter is
encoded using an abstract has_bounds predicate. It is axiomatized for
each array operation. The slice function returns a new array with the
same values inside the slice and the expected bounds. Logical
equality on arrays compares the values in the bounds only. Aggregates
are modified to avoid constraining out-of-bounds values.
On average, this is less efficient than the previous encoding
especially when slices are heavily used. However, this should reduce
the surprise for users when using equality on arrays.