Add tests for GNAT extensions
Add tests that check adequate behavior of GNATprove on GNAT extensions,
which should either reject the code with a clear explanation or accept
it and handle the feature correctly. This is not the case for all
extensions currently, hence some tests are marked XFAIL.

Also slightly adapt a few messages for uniformity.