Use the RAC to create a one liner for counterexamples
This allows printing validated counterexample values. Also use the RAC
to compute the unproved part of the check to avoid discrepencies.
Also fix various minor issues in the RAC that became visible after the
change so as to not decrease the quality of counterexamples in the
testsuite too much.

[changelog]
* src/common/vc_kinds.ads (Cntexmp_Verdict): Add a field for the
environement one-liner and a field for extra information in confirming
verdicts.

* src/counterexamples/ce_display.adb
(Get_Cntexmp_Line_Str): Refactor to share pretty printing between the
display of the environment oneliner and the Why3 counterexamples.
(Get_Environment_One_Liner): Use a custom traversal to accumulate all
relevant counterexample values and reuse the pretty printing used for
Why3 counterexamples.

* src/counterexamples/ce_pretty_printing.adb
(Print_Array_Value): Make printing of array counterexample values
 shorter now array values are displayed more frequently. Fix the
 expected type for the array bounds on empty arrays.
(Print_Discrete): Try to avoid base types introduced by the compiler if
 possible.

* src/counterexamples/ce_rac.ads (Result): Add a field for potential
additional information for RAC_Failure.

* src/counterexamples/ce_rac.adb
(Context): Add a boolean to tell the difference between loop invariant
initialization and loop invariant preservation checks. It is necessary
to avoid validating both VCs at once with the same counterexample
values.
(RAC_Failure): Add an optional parameter for extra information about
the failure.
(Check_Integer): Store the failing bound in the extra information.
(Check_Node): Split assertions to get counterexample values for
top-level quantified variables. Add the part of the assertion which
is failing as extra information.
(Evaluate_Attribute_Prefix_Values): For references to 'Old and
'Loop_Entry, include values for referenced variables to the attribute
maps. They are used when printing counterexample values.
(Find_Binding): Add a flag to support querying values not in the
context.
(Find_Loop_Entry_Value, Find_Old_Value): Factorize queries in the
environement of old and loop_entry values.
(RAC_Attribute_Reference): Use the new query functions for Old and
Loop_Entry and fix handling of Pred and Succ on enumerations as it was
incorrect for characters.
(RAC_Binary_Op): Use the correct type for comparisons.
(RAC_Expr): Add range check on slices.
(RAC_Pragma): Use the First_Loop_Iter flag from the context to decide
whether we are checking the initialization or the preservation of a
loop invariant.
(RAC_Statement): Handle the First_Loop_Iter flag and introduce a scope
for loop in the environment. It is necessary to properly handle nested
loops.

* src/flow/flow_error_messages.adb (Error_Msg_Proof): Use the oneliner
provided by the small step RAC if it has confirmed the verdict.

* src/spark/spark_util.ads (Extra_Info): Export the type for extra
information on unproved checks so it can be shared with the RAC.

* src/why/gnat2why-error_messages.adb
(Extra_Info): Moved to spark_util.ads.
(Compute_Cannot_Prove_Message): Refactored so it can be used for the
RAC.
(VC_Kinds_Match): Loop invariant checks are no longer collapsed with
the other assertions.
(Decide_Cntexmp_Verdict): Create a oneliner from the environement if
the small step RAC confirms the verdict. For now, leave the oneliner
component empty when the verdict is confirmed by the giant step RAC. A
one liner will be computed from the Why3 counterexample instead.
(Handle_Result): If the RAC has found a valid counterexample, it might
not be compatible with the extra information present in the prover
result. Use information computed by the RAC instead.
(Small_Step_Rac): Use the check kind provided by the Why3 result to
match the RAC counterexample. It allows to tell the difference between
loop invariant checks.