Use the RAC to retrieve the VC location and the CE traces
When the RAC verified a check, use it to generate the VC location in
case the VC is split and the traces. Add values at the beginning of the
subprogram and values of loop indexes to the traces when using the RAC
so we do not lose the trace functionality completely.
Also include some minor improvements to the RAC and the CE printing to
avoid losing too many counterexample values in the testsuite. In
particular:
* bounds of scalar types used for range checks are now included in the
  one liner,
* there is a special case for discriminant checks in delta aggregates,
* values of quantified variables are forced to be included in the one
  liner even if they are not mentionned in the check, and
* the detection of internal entities has been adapted to avoid
  excluding entities from inlined subprogram calls.