Support local type invariants
Type invariants are now separated in three classes:
 (1) those which are assumed globally in the unit (they are declared
 in another compilation unit),
 (2) those which can be broken in the current scope, and
 (3) those which are assumed locally in the current scope.
Classes (1) and (2) are handled as before: invariants of class (1) are
included in the dynamic invariant predicate and invariants of class (2)
are handled manually.
Locally assumed type invariants (3) are now added on to of the dynamic
invariant predicate when computing the dynamic invariant. The
distinction between (2) and (3) is based on the scope of the
translation. In most cases, we use the Current_Subp global variable
which holds the entity that is currently translated.

To accomodate more complex scopes, the handling of parameters of
internal subprograms, where some invariants can be relaxed but
potentially not all, is now done in a specific utility function instead
of relying on Is_Globally_Visible.

For internal subprograms declared in other units, it was necessary to
remove the globally assumed invariants (1) as they could be broken.
Now, in this case, the globally assumed invariants are added on to of
the dynamic invariant predicate on a case-by-case basis, like locally
assumed ones (3).