Add information in log file on most difficult proved checks
Compute the list of 10 most difficult checks, as measured by the maximum
time taken to prove any VC which is part of this check. Checks whose max
time is less than one second are not reported. Display the list in the log.

[changelog]
* gnatprove.gpr: Allow GNAT extensions for interpolated strings.
* src/gnatprove/report_database.adb: Update the set of most difficult checks
and only keep the 10 most difficult ones.
* src/gnatprove/report_database.ads: Types to support most difficult checks.
* src/gnatprove/spark_report.adb: Add printing of the most difficult proved
checks in the log. Add clearer headers in the log to separate sections.