Test passes now
The test uses alternate bodies for a unit, but the test actually
hardcodes one body in the project. The other body is in the same source
folder and triggers a hard error (multiple bodies for the same unit).

I took the easy route and removed the unused alternate body. Another
solution could be to restructure the project so that the alternate body
would be invisible from the project.

eng/gpr/gpr-issues#40