Store _is_moved flags in a separate structure
The _is_moved flags used to check reclamation are now stored in a
separate structure declared alongside the object. This structure is
set to moved when the object is moved and reset when the object is
updated.
As this structure does not duplicate information from the object
(arrays bounds, discriminants...) checks for reclamation use both the
move tree and the object.