Prove contracts on Ada.Strings.Unbounded
In order to gain confidence on the newly added contracts on
Ada.Strings.Unbounded, add a test which is fully proved and
has minimal and reviewable differences in terms of spec with
the standard specification and contracts.

The main differences are the following:
- Type Unbounded_String is not controlled, to be in SPARK.
Its Reference component is default-initialized to null instead of
pointing to a static null string, to be allowed in SPARK.
The type of Reference is an access type to Uninit_String, which is
a derived type from String with Relaxed_Initialization, to be able
to catch initialization errors in the code.
It has a predicate stating values for Reference bounds, the relation
between these bounds and component Last, and the range of characters
which should be initialized.
- The spec of Free is marked in SPARK, to be able to analyze its
calls, and it has a contract denoting a deallocation procedure.
- Set_Unbounded_String takes an in out parameter, to be able to
free its Reference. It has also a Depends to reflect that the output
value of Target depends on null.
- Internal procedure Realloc_For_Chunk (in the private part) has a
suitable contract.

The body strunb.adb is based on a-strunb.adb from the stdlib, with
adjustments to be in SPARK (at the exclusion of the bodies of Initialize,
Finalize, Adjust and Free). A lemma from the SPARKlib is used for
proving function "*". Contracts and assertions are added as necessary.
Convertions between String and Uninit_String are added as well as
temporaries to respect ownership rules. Subtypes are added to allow
using "initialized" allocators which in fact use simply the syntax with
qualification, but only to default to an uninitialized value for
an Uninit_String value.

Make that test a replay test, and add corresponding proof sessions.