Support container aggregates on scalar types
To get the aggregate annotation, we search on the base type. This is
not adequate for scalar types whose base type might not be annotated
with the Aggregate aspect. Use the first subtype for scalar types
instead.
Also use the first subtype in the checks for the Aggregate aspect, as
it might be more contraining than the base type.