Detect cases where invariants on incomplete types are ignored
Invariants on incomplete types designated by access types are rejected
when an invariant check is needed (to avoid infinite loops while
computing invariant checks). More generally and for the same reason,
they are ignored when computing the invariants of a value recursively.
This is done for checks (it should not happen in practice as per the
restriction above) and to assume globally assumed invariants on local
subprograms. The second case is not easy to detect in marking and
should only occur when the analysis of some withed units leads to a
violation. Implement a fail safe mechanism raising an error in proof
for this case.