Avoid generating recursive dynamic invariant when possible
Recognize incomplete types whose dynamic invariant is static and use an
axiom on their closed form instead. This is similar to what we do for
static discrete types. For it to be sound, the invariant should be
realizable. Static dynamic invariants are mostly those introduced to
fix the values of missing components in composite types.