Add initial VS Code settings for gnatprove project
These settings allow one to open VS Code from spark2014 repo with
`code .` and get a suitable dev environment for the gnatprove project.
For gnat2why project, this is not done yet, as there are apparently
issued with the Ada plugin to include sources from outside the
workspace, currently checked by the IDE team.

Note that one should answer "no" to the question asked by VS Code:
"Some project source directories are not listed in your workspace:
Do you want to add them?", as answering "yes" changes the current
workspace to an anonymous one with a temporary location for config
files.