Direct definition for static_constraint on access-to-incomplete types
Always give a direct definition to the static_constraint predicate on
access-to-incomplete types instead of using a defining axiom. This
makes it necessary to move the dynamic property predicate used to
fix the designated value of null pointers to the representative theory
of access types (as opposed to their completion theory).