Pull predicates and invariant into proof dependencies at type declaration
Flow would miss a proof dependency when a type is declared but not used in a
unit: proof would still attempt to prove correctness of the unit and needs this
dependency. This is now fixed.

* src/flow/flow-control_flow_graph.adb
Adapt calls to Process_Predicate_And_Invariant to pass the Etype as an actual.
(Do_Type_Declaration): Pull predicate and invariants of the type into proof
dependencies of the analyzed unit.
* src/flow/flow_generated_globals-partial.adb
Adapt calls to Process_Predicate_And_Invariant to pass the Etype as an actual.
* src/flow/flow_utility-proof_dependencies.adb
* src/flow/flow_utility-proof_dependencies.ads
(Process_Predicate_And_Invariant): Change signature. Take a type entity instead
of the node, as only the Etype of the formal was used.
* src/flow/flow_utility.adb
* src/flow/flow_utility.ads
(Pick_Generated_Info_Internal): Adapt calls to Process_Predicate_And_Invariant_Internal.
(Process_Predicate_And_Invariant_Internal): Change signature, adapt implementation.
* testsuite/gnatprove/tests/187__cycle_global/pack.ads
* testsuite/gnatprove/tests/187__cycle_global/test.out
Add a contract in test file. Update output.
* testsuite/gnatprove/tests/247__container_aggregate_capacity/test.out
* testsuite/gnatprove/tests/38__proof_cycle_predicates/test.out
* testsuite/gnatprove/tests/38__proof_cycle_type_invariants/test.out
* testsuite/gnatprove/tests/501__derived_alias/test.out
* testsuite/gnatprove/tests/P725-016__recursive_bad_axiom/test.out
* testsuite/gnatprove/tests/Q520-002__invalidpost/test.out
* testsuite/gnatprove/tests/SA25-033__subprogram_variants/test.out
* testsuite/gnatprove/tests/TU__types_and_subtypes_illegal/test.out
* testsuite/gnatprove/tests/V714-006__multisets/test.out
* testsuite/gnatprove/tests/W111-023__relaxed_init_preds/test.out
Update test output.